# Conflicts:
#	ObfuscationTools/Obfuscation.Core/Obfuscator.cs
#	ObfuscationTools/Obfuscation.Core/bin/Debug/Obfuscation.Core.dll
#	ObfuscationTools/Obfuscation.Core/bin/Debug/Obfuscation.Core.pdb
#	ObfuscationTools/Obfuscation.Core/obj/Debug/Obfuscation.Core.dll
#	ObfuscationTools/Obfuscation.Core/obj/Debug/Obfuscation.Core.pdb
#	ObfuscationTools/UI.Console/bin/Debug/Obfuscation.Core.dll
#	ObfuscationTools/UI.Console/bin/Debug/Obfuscation.Core.pdb
#	ObfuscationTools/UI.Console/bin/Debug/UI.Console.exe
#	ObfuscationTools/UI.Console/bin/Debug/UI.Console.pdb
#	ObfuscationTools/UI.Console/obj/Debug/UI.Console.csprojResolveAssemblyReference.cache
#	ObfuscationTools/UI.Console/obj/Debug/UI.Console.exe
#	ObfuscationTools/UI.Console/obj/Debug/UI.Console.pdb

+++ ObfuscationTools/Obfuscation.Core/ICSharpObfuscator.cs
@@ -14,7 +14,7 @@ namespace Obfuscation.Core
 
         string InlineMethods(string cSharpCode, double inliningCoefficient = 1);
 
-        string UnrollLoops(string cSharpCode, double unrollingCoefficient = 1);
+        string UnrollLoops(string cSharpCode, int unrollingCoefficient = 1);
 
         string InterleaveFuctions(string cSharpCode, int interleavesCount = 1);
